+++ tests/spec/directives/forward/error/load.rs
@@ -51,7 +51,7 @@ mod test_loop {
         );
     }
     #[test]
-    #[ignore] // missing error
+    #[ignore] // wrong error
     fn forward_to_import() {
         let runner = runner().with_cwd("forward_to_import");
         assert_eq!(