+++ tests/spec/directives/import/error/not_found.rs
@@ -14,7 +14,6 @@ fn runner() -> crate::TestRunner {
 }
 
 #[test]
-#[ignore] // missing error
 fn directory_dot_import() {
     let runner = runner().with_cwd("directory_dot_import");
     assert_eq!(
@@ -46,7 +45,6 @@ fn no_extension() {
     );
 }
 #[test]
-#[ignore] // missing error
 fn parent_relative() {
     let runner = runner().with_cwd("parent_relative");
     assert_eq!(