+++ tests/spec/libsass_closed_issues/issue_2106/test.rs
@@ -6,7 +6,6 @@ fn runner() -> crate::TestRunner {
 }
 
 #[test]
-#[ignore] // missing error
 fn test() {
     assert_eq!(
         runner().err("@import \"../does-not-exist\";\n"),